fix: ブックプレビューで動画が自動再生されない
