+++ components/organisms/TopicViewer.stories.tsx
@@ -1,8 +1,14 @@
-export default { title: "organisms/TopicViewer" };
-
+import type { Story } from "@storybook/react";
 import TopicViewer from "./TopicViewer";
 import { topic } from "samples";
 
-const defaultProps = { topic };
+export default { title: "organisms/TopicViewer", component: TopicViewer };
 
-export const Default = () => <TopicViewer {...defaultProps} />;
+const Template: Story<Parameters<typeof TopicViewer>[0]> = (args) => (
+  <TopicViewer {...args} />
+);
+
+export const Default = Template.bind({});
+Default.args = {
+  topic,
+};
