+++ components/organisms/TopicViewer.tsx
@@ -1,9 +1,13 @@
+import { useEffect } from "react";
+import usePrevious from "@rooks/use-previous";
 import clsx from "clsx";
 import { TopicSchema } from "$server/models/topic";
 import Card from "@material-ui/core/Card";
 import { makeStyles } from "@material-ui/core/styles";
 import TopicViewerContent from "$molecules/TopicViewerContent";
 import useCardStyles from "$styles/card";
+import { useBookAtom } from "$store/book";
+import { useVideoAtom } from "$store/video";
 
 const useStyles = makeStyles({
   root: {
@@ -18,10 +22,34 @@ type Props = {
   offset?: number;
 };
 
-export default function TopicViewer(props: Props) {
+export default function TopicViewer({
+  className,
+  topic,
+  onEnded,
+  offset,
+}: Props) {
   const classes = useStyles();
   const cardClasses = useCardStyles();
-  const { className, topic, onEnded, offset } = props;
+  const { book, itemIndex, itemExists } = useBookAtom();
+  const { video, updateVideo } = useVideoAtom();
+  const prevItemIndex = usePrevious(itemIndex);
+  useEffect(() => {
+    if (!book) return;
+    updateVideo(book.sections);
+  }, [book, updateVideo]);
+  useEffect(() => {
+    prevItemIndex &&
+      video.get(itemExists(prevItemIndex)?.resource.url ?? "")?.player.pause();
+    const videoInstance = video.get(itemExists(itemIndex)?.resource.url ?? "");
+    if (!videoInstance) return;
+    if (videoInstance.type == "vimeo") {
+      videoInstance.player.play();
+    } else {
+      videoInstance.player.ready(() => {
+        videoInstance.player.play();
+      });
+    }
+  }, [video, itemExists, prevItemIndex, itemIndex]);
   return (
     <Card classes={cardClasses} className={clsx(classes.root, className)}>
       <TopicViewerContent topic={topic} onEnded={onEnded} offset={offset} />
